feat: Add coupon system

+++ Dtos/User/UpdateUserRequestDTO.cs
@@ -11,6 +11,6 @@ namespace StockMarket.Dtos.User
         public string? Email { get; set; } = string.Empty;
         public string? Password { get; set; } = string.Empty;
         public int? RoleId { get; set; }
-        public int? Balance { get; set; }
+        public decimal? Balance { get; set; }
     }
 }
